Explicitly cast enum value to avoid compilation error on Solaris.
X-Trac-URL: https://ssl.icu-project.org/trac/ticket/7724


git-svn-id: http://source.icu-project.org/repos/icu/icu/trunk@28508 0a8b766c-62c9-45f9-954d-7f2943c9ead0

+++ source/i18n/unicode/dcfmtsym.h
@@ -435,7 +435,7 @@ DecimalFormatSymbols::setSymbol(ENumberFormatSymbol symbol, const UnicodeString 
         if ( u_charDigitValue(sym) == 0 ) {
             for ( int8_t i = 1 ; i<= 9 ; i++ ) {
                 sym++;
-                fSymbols[kOneDigitSymbol+i-1] = UnicodeString(sym);
+                fSymbols[(int)kOneDigitSymbol+i-1] = UnicodeString(sym);
             }
         }
     }
